Handle error from dropsonde/metrics.SendValue
We have chosen to simply implement this error handling without implementing
a test. The sensible way to force an error in an integration test would
be to kill the metron agent listener (or its mock) after dropsonde
initialisation. However, the dropsonde implementation of the emitter
handles disappearing and reappearing listeners (metron agents), so it
didn't work for us.

[#150739352]

Signed-off-by: Jonathan Tsang <jtsang@pivotal.io>